addition CustomSelect and usage
